+++ CHANGELOG.md
@@ -10,12 +10,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `to_tikz` for gridded perms.
 - `to_svg` for gridded perms.
 - `to_gui` for tilings.
-- Mappings for gridded perms: `column_reverse`, `row_complement`, 
+- Mappings for gridded perms: `column_reverse`, `row_complement`,
    `permute_columns`, `permute_rows` and `apply_perm_map_to_cell`.
-- Mappings for tilings: `column_reverse`, `row_complement`, 
+- Mappings for tilings: `column_reverse`, `row_complement`,
    `permute_columns`, `permute_rows`, `apply_perm_map_to_cell`.
 - `guess_from_gridded_perms` to guess tiling T from gridded perms in Grid(T).
 - `enmerate_gp_up_to` counts gridded perms of length 0, 1, 2, ... up to a max length.
+
+### Fixed
+- Anti-diagonal symmetry backward map was fixed.
 
 ## [2.3.1] - 2020-09-11
 ### Fixed